Added UserDetails, Patient and Receptionist entities with relationships

+++ src/main/java/com/example/HospitalManagementSystem/repository/UserDetailsRepository.java
@@ -1,9 +1,7 @@
 package com.example.HospitalManagementSystem.repository;
 
+import com.example.HospitalManagementSystem.model.UserDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.HospitalManagementSystem.model.UserDetails;
-
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
-
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
 }
